1. Intersection methods renamed: * `Set::intersection()` => `Set::intersectionNonStrict()`; * `Set::intersectionStrict()` => `Set::intersection()`; * `Set::partialIntersection()` => `Set::partialIntersectionNonStrict()`; * `Set::partialIntersectionStrict()` => `Set::partialIntersection()`; * `Stream::intersectionWith()` => `Stream::intersectionNonStrictWith()`; * `Stream::intersectionStrictWith()` => `Stream::intersectionWith()`; * `Stream::partialIntersectionWith()` => `Stream::partialIntersectionNonStrictWith()`; * `Stream::partialIntersectionStrictWith()` => `Stream::partialIntersectionWith()`; 2. Tests of these methods renamed. 3. README updated.
