+++ tests/Set/IntersectionTest.php
@@ -60,131 +60,86 @@ class IntersectionTest extends \PHPUnit\Framework\TestCase
             ],
             [
                 [
-                    [null, null],
+                    [1, 2, 3, 4, 5],
+                    [1, 2, 3, 4, 5],
                 ],
-                [null, null],
+                [1, 2, 3, 4, 5],
             ],
             [
                 [
-                    [null, null],
-                    [],
+                    [1, 2, 3, 4, 5],
+                    [1, 2, 3, '4', '5'],
+                ],
+                [1, 2, 3],
+            ],
+            [
+                [
+                    ['1', '2', '3', 4, 5],
+                    [1, 2, 3, '4', '5'],
                 ],
                 [],
             ],
             [
                 [
-                    [null, null, null],
-                    [null, null],
-                    [null],
+                    [1, 1, 2, 2, 1, 1],
+                    [2, 2, 1, 1, 2, 2],
                 ],
-                [null],
+                [2, 1, 2, 1],
             ],
             [
                 [
-                    [1, 2, 3],
-                    [2, 3, 4],
+                    [1, 1, 2, 2, 1, 1],
+                    [2, 2, 1, 1, '2', '2'],
                 ],
-                [2, 3],
+                [2, 1, 2, 1],
             ],
             [
                 [
-                    [1, 2, 3, 4, 5],
-                    [2, 3, 4, 5, 6],
-                    [3, 4, 5, 6, 7],
+                    [1, 1, 2, 2, 1, 1],
+                    [2, 2, '1', '1', 2, 2],
                 ],
-                [3, 4, 5],
+                [2, 2],
             ],
             [
                 [
-                    [2, 3, 4, 5],
-                    [2, 3, 4, 5, 6],
-                    [3, 4, 5, 6, 7],
+                    [1, 1, 1, 1, 1],
+                    [1, 2, 3, 4, 5, 1, 2, 3, 4, 5],
+                    [5, 5, 5, 5, 5, 1, 5, 5, 1],
                 ],
-                [3, 4, 5],
+                [1, 1],
             ],
             [
                 [
-                    [1, 2, 3, 4, 5],
-                    [2, 3, 4, 5, 6],
-                    [3, 4, 5, 6],
+                    [1, 1, 1, 1, 'a'],
+                    [1, 2, 3, 4, 5, 'a', 2, 3, 4, 5],
+                    [5, 5, 5, 5, 5, 'a', 5, 5, 1],
                 ],
-                [3, 4, 5],
+                ['a', 1],
             ],
             [
                 [
-                    [1, 1, 1, 1, 1, 3],
-                    [2, 2, 3, 1, 1],
-                    [3, 1, 1, 3, 3],
+                    [null, 1, 2, 3, 100, null],
+                    [null, 0, 1, 2, 3, 4, null],
+                    [null, -1, 0, 1, 2, 3, 4, 5, null],
+                    [null, -2, -1, 0, 1, 2, 3, 4, 5, 6, null],
+                    [null, -3, -2, -1, 0, 1, 2, 3, 4, 5, 6, 7, null],
                 ],
-                [1, 1, 3],
-            ],
-            [
-                [
-                    [1, 2, 3, 4, 5],
-                    [2, 3, 4, 5, 6, 1],
-                    [3, 4, 5, 6, 1],
-                ],
-                [3, 4, 5, 1],
-            ],
-            [
-                [
-                    [1, 2, 3, 4, 3, 5],
-                    [2, 3, 4, 5, 6, 1, 2, 3],
-                    [3, 4, 5, 6, 1, 2, 3],
-                ],
-                [3, 4, 5, 1, 2, 3],
-            ],
-            [
-                [
-                    [2, 3, 4, 5],
-                    [2, 3, 4, 5, 6],
-                    [3, 4, 5, 6],
-                ],
-                [3, 4, 5],
-            ],
-            [
-                [
-                    [2, 3, 4, 5],
-                    [2, 3, 4, 5, 6],
-                    [],
-                ],
-                [],
-            ],
-            [
-                [
-                    [1.1, 2.1, 3.1],
-                    [2.1, 3.1, 4.1],
-                ],
-                [2.1, 3.1],
-            ],
-            [
-                [
-                    ['1', '2', '3'],
-                    ['2', '3', '4'],
-                ],
-                ['2', '3'],
-            ],
-            [
-                [
-                    ['1', '2', '3', '2', '11'],
-                    ['2', '3', '4', '12', '13', '2'],
-                ],
-                ['2', '3', '2'],
+                [null, 1, 2, 3, null],
             ],
             [
                 [
                     ['1', 2, '3.3', true, false],
                     [true, '2', 3.3, '4', true],
                 ],
-                [true, 2, 3.3, true],
+                [true],
             ],
             [
                 [
-                    ['php', 'python', 'c++', 'java', 'c#', 'javascript', 'typescript'],
-                    ['php', 'python', 'javascript', 'typescript'],
-                    ['php', 'java', 'c#', 'typescript'],
+                    [1, 2, 3, 4, 5, 6, 7, 8, 9],
+                    ['1', '2', 3, 4, 5, 6, 7, '8', '9'],
+                    [1, 3, 5, 7, 9, 11],
                 ],
-                ['php', 'typescript'],
+                [3, 5, 7],
             ],
         ];
     }
@@ -242,131 +197,86 @@ class IntersectionTest extends \PHPUnit\Framework\TestCase
             ],
             [
                 [
-                    $gen([null, null]),
+                    $gen([1, 2, 3, 4, 5]),
+                    $gen([1, 2, 3, 4, 5]),
                 ],
-                [null, null],
+                [1, 2, 3, 4, 5],
             ],
             [
                 [
-                    $gen([null, null]),
-                    $gen([]),
+                    $gen([1, 2, 3, 4, 5]),
+                    $gen([1, 2, 3, '4', '5']),
+                ],
+                [1, 2, 3],
+            ],
+            [
+                [
+                    $gen(['1', '2', '3', 4, 5]),
+                    $gen([1, 2, 3, '4', '5']),
                 ],
                 [],
             ],
             [
                 [
-                    $gen([null, null, null]),
-                    $gen([null, null]),
-                    $gen([null]),
+                    $gen([1, 1, 2, 2, 1, 1]),
+                    $gen([2, 2, 1, 1, 2, 2]),
                 ],
-                [null],
+                [2, 1, 2, 1],
             ],
             [
                 [
-                    $gen([1, 2, 3]),
-                    $gen([2, 3, 4]),
+                    $gen([1, 1, 2, 2, 1, 1]),
+                    $gen([2, 2, 1, 1, '2', '2']),
                 ],
-                [2, 3],
+                [2, 1, 2, 1],
             ],
             [
                 [
-                    $gen([1, 2, 3, 4, 5]),
-                    $gen([2, 3, 4, 5, 6]),
-                    $gen([3, 4, 5, 6, 7]),
+                    $gen([1, 1, 2, 2, 1, 1]),
+                    $gen([2, 2, '1', '1', 2, 2]),
                 ],
-                [3, 4, 5],
+                [2, 2],
             ],
             [
                 [
-                    $gen([2, 3, 4, 5]),
-                    $gen([2, 3, 4, 5, 6]),
-                    $gen([3, 4, 5, 6, 7]),
+                    $gen([1, 1, 1, 1, 1]),
+                    $gen([1, 2, 3, 4, 5, 1, 2, 3, 4, 5]),
+                    $gen([5, 5, 5, 5, 5, 1, 5, 5, 1]),
                 ],
-                [3, 4, 5],
+                [1, 1],
             ],
             [
                 [
-                    $gen([1, 2, 3, 4, 5]),
-                    $gen([2, 3, 4, 5, 6]),
-                    $gen([3, 4, 5, 6]),
+                    $gen([1, 1, 1, 1, 'a']),
+                    $gen([1, 2, 3, 4, 5, 'a', 2, 3, 4, 5]),
+                    $gen([5, 5, 5, 5, 5, 'a', 5, 5, 1]),
                 ],
-                [3, 4, 5],
+                ['a', 1],
             ],
             [
                 [
-                    $gen([1, 1, 1, 1, 1, 3]),
-                    $gen([2, 2, 3, 1, 1]),
-                    $gen([3, 1, 1, 3, 3]),
+                    $gen([null, 1, 2, 3, 100, null]),
+                    $gen([null, 0, 1, 2, 3, 4, null]),
+                    $gen([null, -1, 0, 1, 2, 3, 4, 5, null]),
+                    $gen([null, -2, -1, 0, 1, 2, 3, 4, 5, 6, null]),
+                    $gen([null, -3, -2, -1, 0, 1, 2, 3, 4, 5, 6, 7, null]),
                 ],
-                [1, 1, 3],
-            ],
-            [
-                [
-                    $gen([1, 2, 3, 4, 5]),
-                    $gen([2, 3, 4, 5, 6, 1]),
-                    $gen([3, 4, 5, 6, 1]),
-                ],
-                [3, 4, 5, 1],
-            ],
-            [
-                [
-                    $gen([1, 2, 3, 4, 3, 5]),
-                    $gen([2, 3, 4, 5, 6, 1, 2, 3]),
-                    $gen([3, 4, 5, 6, 1, 2, 3]),
-                ],
-                [3, 4, 5, 1, 2, 3],
-            ],
-            [
-                [
-                    $gen([2, 3, 4, 5]),
-                    $gen([2, 3, 4, 5, 6]),
-                    $gen([3, 4, 5, 6]),
-                ],
-                [3, 4, 5],
-            ],
-            [
-                [
-                    $gen([2, 3, 4, 5]),
-                    $gen([2, 3, 4, 5, 6]),
-                    $gen([]),
-                ],
-                [],
-            ],
-            [
-                [
-                    $gen([1.1, 2.1, 3.1]),
-                    $gen([2.1, 3.1, 4.1]),
-                ],
-                [2.1, 3.1],
-            ],
-            [
-                [
-                    $gen(['1', '2', '3']),
-                    $gen(['2', '3', '4']),
-                ],
-                ['2', '3'],
-            ],
-            [
-                [
-                    $gen(['1', '2', '3', '2', '11']),
-                    $gen(['2', '3', '4', '12', '13', '2']),
-                ],
-                ['2', '3', '2'],
+                [null, 1, 2, 3, null],
             ],
             [
                 [
                     $gen(['1', 2, '3.3', true, false]),
                     $gen([true, '2', 3.3, '4', true]),
                 ],
-                [true, 2, 3.3, true],
+                [true],
             ],
             [
                 [
-                    $gen(['php', 'python', 'c++', 'java', 'c#', 'javascript', 'typescript']),
-                    $gen(['php', 'python', 'javascript', 'typescript']),
-                    $gen(['php', 'java', 'c#', 'typescript']),
+                    $gen([1, 2, 3, 4, 5, 6, 7, 8, 9]),
+                    $gen(['1', '2', 3, 4, 5, 6, 7, '8', '9']),
+                    $gen([1, 3, 5, 7, 9, 11]),
                 ],
-                ['php', 'typescript'],
+                [3, 5, 7],
             ],
         ];
     }
@@ -424,131 +334,86 @@ class IntersectionTest extends \PHPUnit\Framework\TestCase
             ],
             [
                 [
-                    $iter([null, null]),
+                    $iter([1, 2, 3, 4, 5]),
+                    $iter([1, 2, 3, 4, 5]),
                 ],
-                [null, null],
+                [1, 2, 3, 4, 5],
             ],
             [
                 [
-                    $iter([null, null]),
-                    $iter([]),
+                    $iter([1, 2, 3, 4, 5]),
+                    $iter([1, 2, 3, '4', '5']),
+                ],
+                [1, 2, 3],
+            ],
+            [
+                [
+                    $iter(['1', '2', '3', 4, 5]),
+                    $iter([1, 2, 3, '4', '5']),
                 ],
                 [],
             ],
             [
                 [
-                    $iter([null, null, null]),
-                    $iter([null, null]),
-                    $iter([null]),
+                    $iter([1, 1, 2, 2, 1, 1]),
+                    $iter([2, 2, 1, 1, 2, 2]),
                 ],
-                [null],
+                [2, 1, 2, 1],
             ],
             [
                 [
-                    $iter([1, 2, 3]),
-                    $iter([2, 3, 4]),
+                    $iter([1, 1, 2, 2, 1, 1]),
+                    $iter([2, 2, 1, 1, '2', '2']),
                 ],
-                [2, 3],
+                [2, 1, 2, 1],
             ],
             [
                 [
-                    $iter([1, 2, 3, 4, 5]),
-                    $iter([2, 3, 4, 5, 6]),
-                    $iter([3, 4, 5, 6, 7]),
+                    $iter([1, 1, 2, 2, 1, 1]),
+                    $iter([2, 2, '1', '1', 2, 2]),
                 ],
-                [3, 4, 5],
+                [2, 2],
             ],
             [
                 [
-                    $iter([2, 3, 4, 5]),
-                    $iter([2, 3, 4, 5, 6]),
-                    $iter([3, 4, 5, 6, 7]),
+                    $iter([1, 1, 1, 1, 1]),
+                    $iter([1, 2, 3, 4, 5, 1, 2, 3, 4, 5]),
+                    $iter([5, 5, 5, 5, 5, 1, 5, 5, 1]),
                 ],
-                [3, 4, 5],
+                [1, 1],
             ],
             [
                 [
-                    $iter([1, 2, 3, 4, 5]),
-                    $iter([2, 3, 4, 5, 6]),
-                    $iter([3, 4, 5, 6]),
+                    $iter([1, 1, 1, 1, 'a']),
+                    $iter([1, 2, 3, 4, 5, 'a', 2, 3, 4, 5]),
+                    $iter([5, 5, 5, 5, 5, 'a', 5, 5, 1]),
                 ],
-                [3, 4, 5],
+                ['a', 1],
             ],
             [
                 [
-                    $iter([1, 1, 1, 1, 1, 3]),
-                    $iter([2, 2, 3, 1, 1]),
-                    $iter([3, 1, 1, 3, 3]),
+                    $iter([null, 1, 2, 3, 100, null]),
+                    $iter([null, 0, 1, 2, 3, 4, null]),
+                    $iter([null, -1, 0, 1, 2, 3, 4, 5, null]),
+                    $iter([null, -2, -1, 0, 1, 2, 3, 4, 5, 6, null]),
+                    $iter([null, -3, -2, -1, 0, 1, 2, 3, 4, 5, 6, 7, null]),
                 ],
-                [1, 1, 3],
-            ],
-            [
-                [
-                    $iter([1, 2, 3, 4, 5]),
-                    $iter([2, 3, 4, 5, 6, 1]),
-                    $iter([3, 4, 5, 6, 1]),
-                ],
-                [3, 4, 5, 1],
-            ],
-            [
-                [
-                    $iter([1, 2, 3, 4, 3, 5]),
-                    $iter([2, 3, 4, 5, 6, 1, 2, 3]),
-                    $iter([3, 4, 5, 6, 1, 2, 3]),
-                ],
-                [3, 4, 5, 1, 2, 3],
-            ],
-            [
-                [
-                    $iter([2, 3, 4, 5]),
-                    $iter([2, 3, 4, 5, 6]),
-                    $iter([3, 4, 5, 6]),
-                ],
-                [3, 4, 5],
-            ],
-            [
-                [
-                    $iter([2, 3, 4, 5]),
-                    $iter([2, 3, 4, 5, 6]),
-                    $iter([]),
-                ],
-                [],
-            ],
-            [
-                [
-                    $iter([1.1, 2.1, 3.1]),
-                    $iter([2.1, 3.1, 4.1]),
-                ],
-                [2.1, 3.1],
-            ],
-            [
-                [
-                    $iter(['1', '2', '3']),
-                    $iter(['2', '3', '4']),
-                ],
-                ['2', '3'],
-            ],
-            [
-                [
-                    $iter(['1', '2', '3', '2', '11']),
-                    $iter(['2', '3', '4', '12', '13', '2']),
-                ],
-                ['2', '3', '2'],
+                [null, 1, 2, 3, null],
             ],
             [
                 [
                     $iter(['1', 2, '3.3', true, false]),
                     $iter([true, '2', 3.3, '4', true]),
                 ],
-                [true, 2, 3.3, true],
+                [true],
             ],
             [
                 [
-                    $iter(['php', 'python', 'c++', 'java', 'c#', 'javascript', 'typescript']),
-                    $iter(['php', 'python', 'javascript', 'typescript']),
-                    $iter(['php', 'java', 'c#', 'typescript']),
+                    $iter([1, 2, 3, 4, 5, 6, 7, 8, 9]),
+                    $iter(['1', '2', 3, 4, 5, 6, 7, '8', '9']),
+                    $iter([1, 3, 5, 7, 9, 11]),
                 ],
-                ['php', 'typescript'],
+                [3, 5, 7],
             ],
         ];
     }
@@ -606,138 +471,93 @@ class IntersectionTest extends \PHPUnit\Framework\TestCase
             ],
             [
                 [
-                    $trav([null, null]),
+                    $trav([1, 2, 3, 4, 5]),
+                    $trav([1, 2, 3, 4, 5]),
                 ],
-                [null, null],
+                [1, 2, 3, 4, 5],
             ],
             [
                 [
-                    $trav([null, null]),
-                    $trav([]),
+                    $trav([1, 2, 3, 4, 5]),
+                    $trav([1, 2, 3, '4', '5']),
+                ],
+                [1, 2, 3],
+            ],
+            [
+                [
+                    $trav(['1', '2', '3', 4, 5]),
+                    $trav([1, 2, 3, '4', '5']),
                 ],
                 [],
             ],
             [
                 [
-                    $trav([null, null, null]),
-                    $trav([null, null]),
-                    $trav([null]),
+                    $trav([1, 1, 2, 2, 1, 1]),
+                    $trav([2, 2, 1, 1, 2, 2]),
                 ],
-                [null],
+                [2, 1, 2, 1],
             ],
             [
                 [
-                    $trav([1, 2, 3]),
-                    $trav([2, 3, 4]),
+                    $trav([1, 1, 2, 2, 1, 1]),
+                    $trav([2, 2, 1, 1, '2', '2']),
                 ],
-                [2, 3],
+                [2, 1, 2, 1],
             ],
             [
                 [
-                    $trav([1, 2, 3, 4, 5]),
-                    $trav([2, 3, 4, 5, 6]),
-                    $trav([3, 4, 5, 6, 7]),
+                    $trav([1, 1, 2, 2, 1, 1]),
+                    $trav([2, 2, '1', '1', 2, 2]),
                 ],
-                [3, 4, 5],
+                [2, 2],
             ],
             [
                 [
-                    $trav([2, 3, 4, 5]),
-                    $trav([2, 3, 4, 5, 6]),
-                    $trav([3, 4, 5, 6, 7]),
+                    $trav([1, 1, 1, 1, 1]),
+                    $trav([1, 2, 3, 4, 5, 1, 2, 3, 4, 5]),
+                    $trav([5, 5, 5, 5, 5, 1, 5, 5, 1]),
                 ],
-                [3, 4, 5],
+                [1, 1],
             ],
             [
                 [
-                    $trav([1, 2, 3, 4, 5]),
-                    $trav([2, 3, 4, 5, 6]),
-                    $trav([3, 4, 5, 6]),
+                    $trav([1, 1, 1, 1, 'a']),
+                    $trav([1, 2, 3, 4, 5, 'a', 2, 3, 4, 5]),
+                    $trav([5, 5, 5, 5, 5, 'a', 5, 5, 1]),
                 ],
-                [3, 4, 5],
+                ['a', 1],
             ],
             [
                 [
-                    $trav([1, 1, 1, 1, 1, 3]),
-                    $trav([2, 2, 3, 1, 1]),
-                    $trav([3, 1, 1, 3, 3]),
+                    $trav([null, 1, 2, 3, 100, null]),
+                    $trav([null, 0, 1, 2, 3, 4, null]),
+                    $trav([null, -1, 0, 1, 2, 3, 4, 5, null]),
+                    $trav([null, -2, -1, 0, 1, 2, 3, 4, 5, 6, null]),
+                    $trav([null, -3, -2, -1, 0, 1, 2, 3, 4, 5, 6, 7, null]),
                 ],
-                [1, 1, 3],
-            ],
-            [
-                [
-                    $trav([1, 2, 3, 4, 5]),
-                    $trav([2, 3, 4, 5, 6, 1]),
-                    $trav([3, 4, 5, 6, 1]),
-                ],
-                [3, 4, 5, 1],
-            ],
-            [
-                [
-                    $trav([1, 2, 3, 4, 3, 5]),
-                    $trav([2, 3, 4, 5, 6, 1, 2, 3]),
-                    $trav([3, 4, 5, 6, 1, 2, 3]),
-                ],
-                [3, 4, 5, 1, 2, 3],
-            ],
-            [
-                [
-                    $trav([2, 3, 4, 5]),
-                    $trav([2, 3, 4, 5, 6]),
-                    $trav([3, 4, 5, 6]),
-                ],
-                [3, 4, 5],
-            ],
-            [
-                [
-                    $trav([2, 3, 4, 5]),
-                    $trav([2, 3, 4, 5, 6]),
-                    $trav([]),
-                ],
-                [],
-            ],
-            [
-                [
-                    $trav([1.1, 2.1, 3.1]),
-                    $trav([2.1, 3.1, 4.1]),
-                ],
-                [2.1, 3.1],
-            ],
-            [
-                [
-                    $trav(['1', '2', '3']),
-                    $trav(['2', '3', '4']),
-                ],
-                ['2', '3'],
-            ],
-            [
-                [
-                    $trav(['1', '2', '3', '2', '11']),
-                    $trav(['2', '3', '4', '12', '13', '2']),
-                ],
-                ['2', '3', '2'],
+                [null, 1, 2, 3, null],
             ],
             [
                 [
                     $trav(['1', 2, '3.3', true, false]),
                     $trav([true, '2', 3.3, '4', true]),
                 ],
-                [true, 2, 3.3, true],
+                [true],
             ],
             [
                 [
-                    $trav(['php', 'python', 'c++', 'java', 'c#', 'javascript', 'typescript']),
-                    $trav(['php', 'python', 'javascript', 'typescript']),
-                    $trav(['php', 'java', 'c#', 'typescript']),
+                    $trav([1, 2, 3, 4, 5, 6, 7, 8, 9]),
+                    $trav(['1', '2', 3, 4, 5, 6, 7, '8', '9']),
+                    $trav([1, 3, 5, 7, 9, 11]),
                 ],
-                ['php', 'typescript'],
+                [3, 5, 7],
             ],
         ];
     }
 
     /**
      * @dataProvider dataProviderForMixed
-     * @param        array<\iterable> $iterables
+     * @param        array<mixed> $iterables
      * @param        array $expected
      */
     public function testMixed(array $iterables, array $expected): void
@@ -780,39 +600,30 @@ class IntersectionTest extends \PHPUnit\Framework\TestCase
             ],
             [
                 [
-                    [null, null, null, null, null],
-                    $gen([null, null, null, null]),
-                    $iter([null, null, null]),
-                    $trav([null, null]),
+                    [1, 1, 1, 1, 1],
+                    $gen([1, 2, 3, 4, 5, 1, 2, 3, 4, 5]),
+                    $iter([5, 5, 5, 5, 5, 1, 5, 5, 1]),
+                    $trav([]),
                 ],
-                [null, null],
+                [],
             ],
             [
                 [
-                    [1, 2, 3, 4, 5],
-                    $gen([2, 3, 4, 5, 6]),
-                    $iter([3, 4, 5, 6, 7]),
-                    $trav([4, 5, 6, 7, 8]),
+                    [1, 1, 1, 1, 1],
+                    $gen([1, 2, 3, 4, 5, 1, 2, 3, 4, 5]),
+                    $iter([5, 5, 5, 5, 5, 1, 5, 5, 1]),
+                    $trav([5, 2, 5, 1, 1, 2]),
                 ],
-                [4, 5],
+                [1, 1],
             ],
             [
                 [
-                    [1, 2, 3, 4, 5, 1, 2, 3, 4, 5, 1, 2, 3, 4, 5],
-                    $gen([2, 3, 4, 5, 6, 2, 3, 4, 5, 6, 2, 3, 4, 5, 6]),
-                    $iter([3, 4, 5, 6, 7, 3, 4, 5, 6, 7, 3, 4, 5, 6, 7]),
-                    $trav([4, 5, 6, 7, 8, 4, 5, 6, 7, 8, 4, 5, 6, 7, 8]),
+                    [1, 1, 1, 1, '1'],
+                    $gen([1, 2, 3, 4, 5, '1', 2, 3, 4, 5]),
+                    $iter([5, 5, 5, 5, 5, '1', 5, 5, 1]),
+                    $trav([5, 2, 5, 1, 1, 2]),
                 ],
-                [4, 5, 4, 5, 4, 5],
-            ],
-            [
-                [
-                    [1, 2, 3, 4, 5, 1, 2, 3, 4, 5, 1, 2, 3, 4, 5, 11],
-                    $gen([2, 3, 4, 5, 6, 2, 3, 4, 5, 6, 2, 3, 4, 5, 6, 12, 13]),
-                    $iter([3, 4, 5, 6, 7, 3, 4, 5, 6, 7, 3, 4, 5, 6, 7, 14, 15, 16]),
-                    $trav([4, 5, 6, 7, 8, 4, 5, 6, 7, 8, 4, 5, 6, 7, 8, 17, 18, 19, 20]),
-                ],
-                [4, 5, 4, 5, 4, 5],
+                [1],
             ],
         ];
     }
