+++ tests/Set/PartialIntersectionTest.php
@@ -84,161 +84,71 @@ class PartialIntersectionTest extends \PHPUnit\Framework\TestCase
             ],
             [
                 [
-                    [],
-                    [],
-                    [],
-                ],
-                1,
-                [],
-            ],
-            [
-                [
-                    [],
-                    [],
-                    [],
+                    [1, 2, 3, 4, 5],
+                    [1, 2, 3, 4, 5],
                 ],
                 2,
-                [],
-            ],
-            [
-                [
-                    [],
-                    [],
-                    [],
-                ],
-                3,
-                [],
-            ],
-            [
-                [
-                    [],
-                    [],
-                    [],
-                ],
-                4,
-                [],
-            ],
-            [
-                [
-                    [1, 2, 3],
-                ],
-                1,
-                [1, 2, 3],
-            ],
-            [
-                [
-                    [1, 2, 3],
-                ],
-                2,
-                [],
-            ],
-            [
-                [
-                    [1, 2, 3],
-                    [2, 3, 4, 5],
-                ],
-                1,
                 [1, 2, 3, 4, 5],
             ],
             [
                 [
-                    [1, 2, 3],
-                    [2, 3, 4, 5],
+                    [1, 2, 3, 4, 5],
+                    [1, 2, '3', '4', 5],
                 ],
                 2,
-                [2, 3],
+                [1, 2, 5],
             ],
             [
                 [
-                    [1, 2, 3],
-                    [2, 3, 4, 5],
+                    ['1', '2', 3, 4, 5],
+                    ['1', 2, '3', '4', 5],
+                ],
+                2,
+                ['1', 5],
+            ],
+            [
+                [
+                    [1, 1, 1, 1, 1],
+                    [1, 2, 3, 4, 5, 6],
+                    [1, 5, 5, 5, 5, 5, 5],
+                ],
+                2,
+                [1, 5],
+            ],
+            [
+                [
+                    [1, 1, 1, 1, 1],
+                    [1, 2, 3, 4, '5', 6],
+                    [1, 5, 5, 5, 5, 5, 5],
+                ],
+                2,
+                [1],
+            ],
+            [
+                [
+                    [1, 1, 1, 1, 1],
+                    [1, 2, 3, 4, 5, 6],
+                    [1, 5, 5, 5, 5, 5, 5],
                 ],
                 3,
-                [],
+                [1],
             ],
             [
                 [
-                    [1, 2, 3, 2],
-                    [2, 3, 4, 5, 2, 1],
-                ],
-                1,
-                [1, 2, 3, 4, 2, 5],
-            ],
-            [
-                [
-                    [1, 2, 3, 2],
-                    [2, 3, 4, 5, 2, 1],
+                    [1, true, '1', 1.0, '1.0'],
+                    [true, true, true, true, true, true],
                 ],
                 2,
-                [2, 3, 2, 1],
+                [true],
             ],
             [
                 [
-                    [1, 2, 3, 2],
-                    [2, 3, 4, 5, 2, 1],
-                ],
-                3,
-                [],
-            ],
-            [
-                [
-                    [1, 2, 3, 2],
-                    [2, 3, 4, 5, 2, 1],
-                    [1, 2, 3],
-                ],
-                1,
-                [1, 2, 3, 4, 2, 5],
-            ],
-            [
-                [
-                    [1, 2, 3, 2],
-                    [2, 3, 4, 5, 2, 1],
-                    [1, 2, 3],
+                    [1, 2, 3, 4, 5, 6, 7, 8, 9],
+                    ['1', '2', 3, 4, 5, 6, 7, '8', '9'],
+                    [1, 3, 5, 7, 9, 11],
                 ],
                 2,
-                [1, 2, 3, 2],
-            ],
-            [
-                [
-                    [1, 2, 3, 2],
-                    [2, 3, 4, 5, 2, 1],
-                    [1, 2, 3],
-                ],
-                3,
-                [2, 3, 1],
-            ],
-            [
-                [
-                    [null, null],
-                    [null, null, null],
-                ],
-                1,
-                [null, null, null],
-            ],
-            [
-                [
-                    [null, null],
-                    [null, null, null],
-                ],
-                2,
-                [null, null],
-            ],
-            [
-                [
-                    [null, null],
-                    [null, null, null],
-                ],
-                3,
-                [],
-            ],
-            [
-                [
-                    ['php', 'python', 'c++', 'java', 'c#', 'javascript', 'typescript'],
-                    ['php', 'python', 'javascript', 'typescript'],
-                    ['php', 'java', 'c#', 'typescript'],
-                ],
-                2,
-                ['php', 'python', 'java', 'typescript', 'c#', 'javascript'],
+                [1, 3, 4, 5, 6, 7, 9],
             ],
         ];
     }
@@ -320,161 +230,71 @@ class PartialIntersectionTest extends \PHPUnit\Framework\TestCase
             ],
             [
                 [
-                    $gen([]),
-                    $gen([]),
-                    $gen([]),
-                ],
-                1,
-                [],
-            ],
-            [
-                [
-                    $gen([]),
-                    $gen([]),
-                    $gen([]),
+                    $gen([1, 2, 3, 4, 5]),
+                    $gen([1, 2, 3, 4, 5]),
                 ],
                 2,
-                [],
-            ],
-            [
-                [
-                    $gen([]),
-                    $gen([]),
-                    $gen([]),
-                ],
-                3,
-                [],
-            ],
-            [
-                [
-                    $gen([]),
-                    $gen([]),
-                    $gen([]),
-                ],
-                4,
-                [],
-            ],
-            [
-                [
-                    $gen([1, 2, 3]),
-                ],
-                1,
-                [1, 2, 3],
-            ],
-            [
-                [
-                    $gen([1, 2, 3]),
-                ],
-                2,
-                [],
-            ],
-            [
-                [
-                    $gen([1, 2, 3]),
-                    $gen([2, 3, 4, 5]),
-                ],
-                1,
                 [1, 2, 3, 4, 5],
             ],
             [
                 [
-                    $gen([1, 2, 3]),
-                    $gen([2, 3, 4, 5]),
+                    $gen([1, 2, 3, 4, 5]),
+                    $gen([1, 2, '3', '4', 5]),
                 ],
                 2,
-                [2, 3],
+                [1, 2, 5],
             ],
             [
                 [
-                    $gen([1, 2, 3]),
-                    $gen([2, 3, 4, 5]),
+                    $gen(['1', '2', 3, 4, 5]),
+                    $gen(['1', 2, '3', '4', 5]),
+                ],
+                2,
+                ['1', 5],
+            ],
+            [
+                [
+                    $gen([1, 1, 1, 1, 1]),
+                    $gen([1, 2, 3, 4, 5, 6]),
+                    $gen([1, 5, 5, 5, 5, 5, 5]),
+                ],
+                2,
+                [1, 5],
+            ],
+            [
+                [
+                    $gen([1, 1, 1, 1, 1]),
+                    $gen([1, 2, 3, 4, '5', 6]),
+                    $gen([1, 5, 5, 5, 5, 5, 5]),
+                ],
+                2,
+                [1],
+            ],
+            [
+                [
+                    $gen([1, 1, 1, 1, 1]),
+                    $gen([1, 2, 3, 4, 5, 6]),
+                    $gen([1, 5, 5, 5, 5, 5, 5]),
                 ],
                 3,
-                [],
+                [1],
             ],
             [
                 [
-                    $gen([1, 2, 3, 2]),
-                    $gen([2, 3, 4, 5, 2, 1]),
-                ],
-                1,
-                [1, 2, 3, 4, 2, 5],
-            ],
-            [
-                [
-                    $gen([1, 2, 3, 2]),
-                    $gen([2, 3, 4, 5, 2, 1]),
+                    $gen([1, true, '1', 1.0, '1.0']),
+                    $gen([true, true, true, true, true, true]),
                 ],
                 2,
-                [2, 3, 2, 1],
+                [true],
             ],
             [
                 [
-                    $gen([1, 2, 3, 2]),
-                    $gen([2, 3, 4, 5, 2, 1]),
-                ],
-                3,
-                [],
-            ],
-            [
-                [
-                    $gen([1, 2, 3, 2]),
-                    $gen([2, 3, 4, 5, 2, 1]),
-                    $gen([1, 2, 3]),
-                ],
-                1,
-                [1, 2, 3, 4, 2, 5],
-            ],
-            [
-                [
-                    $gen([1, 2, 3, 2]),
-                    $gen([2, 3, 4, 5, 2, 1]),
-                    $gen([1, 2, 3]),
+                    $gen([1, 2, 3, 4, 5, 6, 7, 8, 9]),
+                    $gen(['1', '2', 3, 4, 5, 6, 7, '8', '9']),
+                    $gen([1, 3, 5, 7, 9, 11]),
                 ],
                 2,
-                [1, 2, 3, 2],
-            ],
-            [
-                [
-                    $gen([1, 2, 3, 2]),
-                    $gen([2, 3, 4, 5, 2, 1]),
-                    $gen([1, 2, 3]),
-                ],
-                3,
-                [2, 3, 1],
-            ],
-            [
-                [
-                    $gen([null, null]),
-                    $gen([null, null, null]),
-                ],
-                1,
-                [null, null, null],
-            ],
-            [
-                [
-                    $gen([null, null]),
-                    $gen([null, null, null]),
-                ],
-                2,
-                [null, null],
-            ],
-            [
-                [
-                    $gen([null, null]),
-                    $gen([null, null, null]),
-                ],
-                3,
-                [],
-            ],
-            [
-                [
-                    $gen(['php', 'python', 'c++', 'java', 'c#', 'javascript', 'typescript']),
-                    $gen(['php', 'python', 'javascript', 'typescript']),
-                    $gen(['php', 'java', 'c#', 'typescript']),
-                ],
-                2,
-                ['php', 'python', 'java', 'typescript', 'c#', 'javascript'],
+                [1, 3, 4, 5, 6, 7, 9],
             ],
         ];
     }
@@ -556,161 +376,71 @@ class PartialIntersectionTest extends \PHPUnit\Framework\TestCase
             ],
             [
                 [
-                    $iter([]),
-                    $iter([]),
-                    $iter([]),
-                ],
-                1,
-                [],
-            ],
-            [
-                [
-                    $iter([]),
-                    $iter([]),
-                    $iter([]),
+                    $iter([1, 2, 3, 4, 5]),
+                    $iter([1, 2, 3, 4, 5]),
                 ],
                 2,
-                [],
-            ],
-            [
-                [
-                    $iter([]),
-                    $iter([]),
-                    $iter([]),
-                ],
-                3,
-                [],
-            ],
-            [
-                [
-                    $iter([]),
-                    $iter([]),
-                    $iter([]),
-                ],
-                4,
-                [],
-            ],
-            [
-                [
-                    $iter([1, 2, 3]),
-                ],
-                1,
-                [1, 2, 3],
-            ],
-            [
-                [
-                    $iter([1, 2, 3]),
-                ],
-                2,
-                [],
-            ],
-            [
-                [
-                    $iter([1, 2, 3]),
-                    $iter([2, 3, 4, 5]),
-                ],
-                1,
                 [1, 2, 3, 4, 5],
             ],
             [
                 [
-                    $iter([1, 2, 3]),
-                    $iter([2, 3, 4, 5]),
+                    $iter([1, 2, 3, 4, 5]),
+                    $iter([1, 2, '3', '4', 5]),
                 ],
                 2,
-                [2, 3],
+                [1, 2, 5],
             ],
             [
                 [
-                    $iter([1, 2, 3]),
-                    $iter([2, 3, 4, 5]),
+                    $iter(['1', '2', 3, 4, 5]),
+                    $iter(['1', 2, '3', '4', 5]),
+                ],
+                2,
+                ['1', 5],
+            ],
+            [
+                [
+                    $iter([1, 1, 1, 1, 1]),
+                    $iter([1, 2, 3, 4, 5, 6]),
+                    $iter([1, 5, 5, 5, 5, 5, 5]),
+                ],
+                2,
+                [1, 5],
+            ],
+            [
+                [
+                    $iter([1, 1, 1, 1, 1]),
+                    $iter([1, 2, 3, 4, '5', 6]),
+                    $iter([1, 5, 5, 5, 5, 5, 5]),
+                ],
+                2,
+                [1],
+            ],
+            [
+                [
+                    $iter([1, 1, 1, 1, 1]),
+                    $iter([1, 2, 3, 4, 5, 6]),
+                    $iter([1, 5, 5, 5, 5, 5, 5]),
                 ],
                 3,
-                [],
+                [1],
             ],
             [
                 [
-                    $iter([1, 2, 3, 2]),
-                    $iter([2, 3, 4, 5, 2, 1]),
-                ],
-                1,
-                [1, 2, 3, 4, 2, 5],
-            ],
-            [
-                [
-                    $iter([1, 2, 3, 2]),
-                    $iter([2, 3, 4, 5, 2, 1]),
+                    $iter([1, true, '1', 1.0, '1.0']),
+                    $iter([true, true, true, true, true, true]),
                 ],
                 2,
-                [2, 3, 2, 1],
+                [true],
             ],
             [
                 [
-                    $iter([1, 2, 3, 2]),
-                    $iter([2, 3, 4, 5, 2, 1]),
-                ],
-                3,
-                [],
-            ],
-            [
-                [
-                    $iter([1, 2, 3, 2]),
-                    $iter([2, 3, 4, 5, 2, 1]),
-                    $iter([1, 2, 3]),
-                ],
-                1,
-                [1, 2, 3, 4, 2, 5],
-            ],
-            [
-                [
-                    $iter([1, 2, 3, 2]),
-                    $iter([2, 3, 4, 5, 2, 1]),
-                    $iter([1, 2, 3]),
+                    $iter([1, 2, 3, 4, 5, 6, 7, 8, 9]),
+                    $iter(['1', '2', 3, 4, 5, 6, 7, '8', '9']),
+                    $iter([1, 3, 5, 7, 9, 11]),
                 ],
                 2,
-                [1, 2, 3, 2],
-            ],
-            [
-                [
-                    $iter([1, 2, 3, 2]),
-                    $iter([2, 3, 4, 5, 2, 1]),
-                    $iter([1, 2, 3]),
-                ],
-                3,
-                [2, 3, 1],
-            ],
-            [
-                [
-                    $iter([null, null]),
-                    $iter([null, null, null]),
-                ],
-                1,
-                [null, null, null],
-            ],
-            [
-                [
-                    $iter([null, null]),
-                    $iter([null, null, null]),
-                ],
-                2,
-                [null, null],
-            ],
-            [
-                [
-                    $iter([null, null]),
-                    $iter([null, null, null]),
-                ],
-                3,
-                [],
-            ],
-            [
-                [
-                    $iter(['php', 'python', 'c++', 'java', 'c#', 'javascript', 'typescript']),
-                    $iter(['php', 'python', 'javascript', 'typescript']),
-                    $iter(['php', 'java', 'c#', 'typescript']),
-                ],
-                2,
-                ['php', 'python', 'java', 'typescript', 'c#', 'javascript'],
+                [1, 3, 4, 5, 6, 7, 9],
             ],
         ];
     }
@@ -792,168 +522,78 @@ class PartialIntersectionTest extends \PHPUnit\Framework\TestCase
             ],
             [
                 [
-                    $trav([]),
-                    $trav([]),
-                    $trav([]),
-                ],
-                1,
-                [],
-            ],
-            [
-                [
-                    $trav([]),
-                    $trav([]),
-                    $trav([]),
+                    $trav([1, 2, 3, 4, 5]),
+                    $trav([1, 2, 3, 4, 5]),
                 ],
                 2,
-                [],
-            ],
-            [
-                [
-                    $trav([]),
-                    $trav([]),
-                    $trav([]),
-                ],
-                3,
-                [],
-            ],
-            [
-                [
-                    $trav([]),
-                    $trav([]),
-                    $trav([]),
-                ],
-                4,
-                [],
-            ],
-            [
-                [
-                    $trav([1, 2, 3]),
-                ],
-                1,
-                [1, 2, 3],
-            ],
-            [
-                [
-                    $trav([1, 2, 3]),
-                ],
-                2,
-                [],
-            ],
-            [
-                [
-                    $trav([1, 2, 3]),
-                    $trav([2, 3, 4, 5]),
-                ],
-                1,
                 [1, 2, 3, 4, 5],
             ],
             [
                 [
-                    $trav([1, 2, 3]),
-                    $trav([2, 3, 4, 5]),
+                    $trav([1, 2, 3, 4, 5]),
+                    $trav([1, 2, '3', '4', 5]),
                 ],
                 2,
-                [2, 3],
+                [1, 2, 5],
             ],
             [
                 [
-                    $trav([1, 2, 3]),
-                    $trav([2, 3, 4, 5]),
+                    $trav(['1', '2', 3, 4, 5]),
+                    $trav(['1', 2, '3', '4', 5]),
+                ],
+                2,
+                ['1', 5],
+            ],
+            [
+                [
+                    $trav([1, 1, 1, 1, 1]),
+                    $trav([1, 2, 3, 4, 5, 6]),
+                    $trav([1, 5, 5, 5, 5, 5, 5]),
+                ],
+                2,
+                [1, 5],
+            ],
+            [
+                [
+                    $trav([1, 1, 1, 1, 1]),
+                    $trav([1, 2, 3, 4, '5', 6]),
+                    $trav([1, 5, 5, 5, 5, 5, 5]),
+                ],
+                2,
+                [1],
+            ],
+            [
+                [
+                    $trav([1, 1, 1, 1, 1]),
+                    $trav([1, 2, 3, 4, 5, 6]),
+                    $trav([1, 5, 5, 5, 5, 5, 5]),
                 ],
                 3,
-                [],
+                [1],
             ],
             [
                 [
-                    $trav([1, 2, 3, 2]),
-                    $trav([2, 3, 4, 5, 2, 1]),
-                ],
-                1,
-                [1, 2, 3, 4, 2, 5],
-            ],
-            [
-                [
-                    $trav([1, 2, 3, 2]),
-                    $trav([2, 3, 4, 5, 2, 1]),
+                    $trav([1, true, '1', 1.0, '1.0']),
+                    $trav([true, true, true, true, true, true]),
                 ],
                 2,
-                [2, 3, 2, 1],
+                [true],
             ],
             [
                 [
-                    $trav([1, 2, 3, 2]),
-                    $trav([2, 3, 4, 5, 2, 1]),
-                ],
-                3,
-                [],
-            ],
-            [
-                [
-                    $trav([1, 2, 3, 2]),
-                    $trav([2, 3, 4, 5, 2, 1]),
-                    $trav([1, 2, 3]),
-                ],
-                1,
-                [1, 2, 3, 4, 2, 5],
-            ],
-            [
-                [
-                    $trav([1, 2, 3, 2]),
-                    $trav([2, 3, 4, 5, 2, 1]),
-                    $trav([1, 2, 3]),
+                    $trav([1, 2, 3, 4, 5, 6, 7, 8, 9]),
+                    $trav(['1', '2', 3, 4, 5, 6, 7, '8', '9']),
+                    $trav([1, 3, 5, 7, 9, 11]),
                 ],
                 2,
-                [1, 2, 3, 2],
-            ],
-            [
-                [
-                    $trav([1, 2, 3, 2]),
-                    $trav([2, 3, 4, 5, 2, 1]),
-                    $trav([1, 2, 3]),
-                ],
-                3,
-                [2, 3, 1],
-            ],
-            [
-                [
-                    $trav([null, null]),
-                    $trav([null, null, null]),
-                ],
-                1,
-                [null, null, null],
-            ],
-            [
-                [
-                    $trav([null, null]),
-                    $trav([null, null, null]),
-                ],
-                2,
-                [null, null],
-            ],
-            [
-                [
-                    $trav([null, null]),
-                    $trav([null, null, null]),
-                ],
-                3,
-                [],
-            ],
-            [
-                [
-                    $trav(['php', 'python', 'c++', 'java', 'c#', 'javascript', 'typescript']),
-                    $trav(['php', 'python', 'javascript', 'typescript']),
-                    $trav(['php', 'java', 'c#', 'typescript']),
-                ],
-                2,
-                ['php', 'python', 'java', 'typescript', 'c#', 'javascript'],
+                [1, 3, 4, 5, 6, 7, 9],
             ],
         ];
     }
 
     /**
      * @dataProvider dataProviderForMixed
-     * @param array<iterable> $iterables
+     * @param array<mixed> $iterables
      * @param positive-int $minIntersectionCount
      * @param array $expected
      */
@@ -1038,293 +678,63 @@ class PartialIntersectionTest extends \PHPUnit\Framework\TestCase
             ],
             [
                 [
-                    [null, null],
-                    $gen([null, 1, null]),
-                    $iter([null, 2, 3, null]),
-                    $trav([null, 2, 4, 3, null]),
+                    [1, 5],
+                    $gen([2, 6]),
+                    $iter([3, 7]),
+                    $trav([4]),
                 ],
                 1,
-                [null, null, 1, 2, 3, 4],
+                [1, 2, 3, 4, 5, 6, 7],
             ],
             [
                 [
-                    [null, null],
-                    $gen([null, 1, null]),
-                    $iter([null, 2, 3, null]),
-                    $trav([null, 2, 4, 3, null]),
+                    [1, 5],
+                    $gen([2, 6]),
+                    $iter([3, 7]),
+                    $trav([4]),
                 ],
                 2,
-                [null, 2, null, 3],
-            ],
-            [
-                [
-                    [null, null],
-                    $gen([null, 1, null]),
-                    $iter([null, 2, 3, null]),
-                    $trav([null, 2, 4, 3, null]),
-                ],
-                3,
-                [null, null],
-            ],
-            [
-                [
-                    [null, null],
-                    $gen([null, 1, null]),
-                    $iter([null, 2, 3, null]),
-                    $trav([null, 2, 4, 3, null]),
-                ],
-                4,
-                [null, null],
-            ],
-            [
-                [
-                    [null, null],
-                    $gen([null, 1, null]),
-                    $iter([null, 2, 3, null]),
-                    $trav([null, 2, 4, 3, null]),
-                ],
-                5,
                 [],
             ],
             [
                 [
-                    [null, null],
-                    $gen([null, null, null]),
-                    $iter([null, null, null, null]),
-                    $trav([null, null, null, null, null]),
-                ],
-                1,
-                [null, null, null, null, null],
-            ],
-            [
-                [
-                    [null, null],
-                    $gen([null, null, null]),
-                    $iter([null, null, null, null]),
-                    $trav([null, null, null, null, null]),
+                    [1, 1, 1, 1, 1],
+                    $gen(['1', true, '1.0', 1, 1.0, 'a', 'b', 'c']),
+                    $iter([1, '1', 1, '1']),
+                    $trav([1, true, 1]),
                 ],
                 2,
-                [null, null, null, null],
+                [1, '1', true, 1],
             ],
             [
                 [
-                    [null, null],
-                    $gen([null, null, null]),
-                    $iter([null, null, null, null]),
-                    $trav([null, null, null, null, null]),
+                    [1, 1, 1, 1, 1],
+                    $gen(['1', true, '1.0', 1, 1.0, 'a', 'b', 'c']),
+                    $iter([1, '1', 1, '1']),
+                    $trav([1, true, 1]),
                 ],
                 3,
-                [null, null, null],
+                [1, 1],
             ],
             [
                 [
-                    [null, null],
-                    $gen([null, null, null]),
-                    $iter([null, null, null, null]),
-                    $trav([null, null, null, null, null]),
+                    [1, 1, 1, 1, 1],
+                    $gen(['1', true, '1.0', 1, 1.0, 'a', 'b', 'c']),
+                    $iter([1, '1', 1, '1']),
+                    $trav([1, true, 1]),
                 ],
                 4,
-                [null, null],
+                [1],
             ],
             [
                 [
-                    [null, null],
-                    $gen([null, null, null]),
-                    $iter([null, null, null, null]),
-                    $trav([null, null, null, null, null]),
+                    [1, 1, 1, 1, 1],
+                    $gen(['1', true, '1.0', 1, 1.0, 'a', 'b', 'c']),
+                    $iter([1, '1', 1, '1']),
+                    $trav([1, true, 1]),
                 ],
                 5,
                 [],
-            ],
-            [
-                [
-                    [1, 2],
-                    $gen([1, 2, 3]),
-                    $iter([1, 2, 3, 4]),
-                    $trav([1, 2, 3, 4, 5]),
-                ],
-                1,
-                [1, 2, 3, 4, 5],
-            ],
-            [
-                [
-                    [1, 2],
-                    $gen([1, 2, 3]),
-                    $iter([1, 2, 3, 4]),
-                    $trav([1, 2, 3, 4, 5]),
-                ],
-                2,
-                [1, 2, 3, 4],
-            ],
-            [
-                [
-                    [1, 2],
-                    $gen([1, 2, 3]),
-                    $iter([1, 2, 3, 4]),
-                    $trav([1, 2, 3, 4, 5]),
-                ],
-                3,
-                [1, 2, 3],
-            ],
-            [
-                [
-                    [1, 2],
-                    $gen([1, 2, 3]),
-                    $iter([1, 2, 3, 4]),
-                    $trav([1, 2, 3, 4, 5]),
-                ],
-                4,
-                [1, 2],
-            ],
-            [
-                [
-                    [1, 2],
-                    $gen([1, 2, 3]),
-                    $iter([1, 2, 3, 4]),
-                    $trav([1, 2, 3, 4, 5]),
-                ],
-                5,
-                [],
-            ],
-            [
-                [
-                    [1, 2, 1, 2],
-                    $gen([1, 2, 3, 1]),
-                    $iter([1, 2, 3, 4]),
-                    $trav([1, 2, 3, 4]),
-                ],
-                1,
-                [1, 2, 1, 3, 2, 4],
-            ],
-            [
-                [
-                    [1, 2, 1, 2],
-                    $gen([1, 2, 3, 1]),
-                    $iter([1, 2, 3, 4]),
-                    $trav([1, 2, 3, 4]),
-                ],
-                2,
-                [1, 2, 3, 1, 4],
-            ],
-            [
-                [
-                    [1, 2, 1, 2],
-                    $gen([1, 2, 3, 1]),
-                    $iter([1, 2, 3, 4]),
-                    $trav([1, 2, 3, 4]),
-                ],
-                3,
-                [1, 2, 3],
-            ],
-            [
-                [
-                    [1, 2, 1, 2],
-                    $gen([1, 2, 3, 1]),
-                    $iter([1, 2, 3, 4]),
-                    $trav([1, 2, 3, 4]),
-                ],
-                4,
-                [1, 2],
-            ],
-            [
-                [
-                    [1, 2, 1, 2],
-                    $gen([1, 2, 3, 1]),
-                    $iter([1, 2, 3, 4]),
-                    $trav([1, 2, 3, 4]),
-                ],
-                5,
-                [],
-            ],
-            [
-                [
-                    [5, 6, 7, 8],
-                    $gen([4, 5, 6, 7]),
-                    $iter([3, 4, 5, 6]),
-                    $trav([2, 3, 4, 5]),
-                ],
-                1,
-                [5, 4, 3, 2, 6, 7, 8],
-            ],
-            [
-                [
-                    [5, 6, 7, 8],
-                    $gen([4, 5, 6, 7, 4]),
-                    $iter([3, 4, 5, 6, 3, 4]),
-                    $trav([2, 3, 4, 5, 2, 3, 4]),
-                ],
-                1,
-                [5, 4, 3, 2, 6, 7, 8, 4, 3, 2],
-            ],
-            [
-                [
-                    [1, 2, 1, 2],
-                    $gen([1, 2, 3, 1, 2, 3]),
-                    $iter([1, 2, 3, 4, 1, 2, 3, 4]),
-                    $trav([1, 2, 3, 4, 5, 1, 2, 3, 4, 5]),
-                ],
-                1,
-                [1, 2, 1, 3, 2, 4, 5, 3, 4, 5],
-            ],
-            [
-                [
-                    [1, 2, 1, 2],
-                    $gen([1, 2, 1, 2]),
-                    $iter([1, 2, 1, 2]),
-                    $trav([1, 2, 1, 2]),
-                ],
-                1,
-                [1, 2, 1, 2],
-            ],
-            [
-                [
-                    [1, 2, 1, 2],
-                    $gen([1, 2, 1, 2]),
-                    $iter([1, 2, 1, 2]),
-                    $trav([1, 2, 1, 2]),
-                ],
-                2,
-                [1, 2, 1, 2],
-            ],
-            [
-                [
-                    [1, 2, 1, 2],
-                    $gen([1, 2, 1, 2]),
-                    $iter([1, 2, 1, 2]),
-                    $trav([1, 2, 1, 2]),
-                ],
-                3,
-                [1, 2, 1, 2],
-            ],
-            [
-                [
-                    [1, 2, 1, 2],
-                    $gen([1, 2, 1, 2]),
-                    $iter([1, 2, 1, 2]),
-                    $trav([1, 2, 1, 2]),
-                ],
-                4,
-                [1, 2, 1, 2],
-            ],
-            [
-                [
-                    [1, 2, 1, 2],
-                    $gen([1, 2, 1, 2]),
-                    $iter([1, 2, 1, 2]),
-                    $trav([1, 2, 1, 2]),
-                ],
-                5,
-                [],
-            ],
-            [
-                [
-                    [1, 2, 1, 2],
-                    $gen([1, 2, 1, 2, 1, 2]),
-                    $iter([1, 2, 1, 2, 1, 2, 1, 2]),
-                    $trav([1, 2, 1, 2, 1, 2, 1, 2, 1, 2]),
-                ],
-                1,
-                [1, 2, 1, 2, 1, 2, 1, 2, 1, 2],
             ],
         ];
     }
